Improve contributing
Just after the installation, ionic show me a white page, and I find some white papers. It was late and I was going to sleep.
After a good night, I see a gulpfile and try to use gulp. Youra, It's working! So, to help another developers, I improve the doc :)